+++ libnavui-maps/src/main/java/com/mapbox/navigation/ui/maps/route/line/model/RouteSetValue.kt
@@ -58,3 +58,17 @@ fun interface RouteLineExpressionProvider {
      */
     fun generateExpression(): Expression
 }
+
+/**
+ * Represents a function that returns an [Expression]. The expression this provider is expected
+ * to produce is a lineTrimOffset expression. This is a specific type of expression that will
+ * make a line transparent between two values representing sections of a line.
+ *
+ * For example a call like literal(listOf(0.0, 0.5)) would produce a trim offset expression
+ * that made a line transparent from the beginning of the line to the midpoint of the line. In other
+ * words the first 50% of the line would be transparent.  A call like literal(listOf(0.25, 0.5))
+ * would make the line transparent starting at 25% of the line's length to 50% of the line's length.
+ * The line's color would be represented in the other sections of the line.  See the Map API documentation
+ * regarding lineTrimOffset for more information.
+ */
+fun interface RouteLineTrimExpressionProvider : RouteLineExpressionProvider